fixed cluster mode

+++ db/providers/redis/users.js
@@ -7,6 +7,8 @@ const UserIndexKey = "user.n";
 const UserKey = "user.";
 
 function Users(){
+	if (!!process.env.CLUSTER_MODE)
+		return;
 	var client = redis.createClient(options);
 	client.del(UserIndexKey); // overwrite user data by always starting from userId=1
